+++ src/main/php/lang.base.php
@@ -450,12 +450,14 @@ function is($type, $object) {
     return is_bool($object);
   } else if ('var' === $type) {
     return true;
+  } else if (0 === strncmp($type, 'function(', 9)) {
+    return \lang\FunctionType::forName($type)->isInstance($object);
   } else if (0 === substr_compare($type, '[]', -2)) {
     return (new \lang\ArrayType(substr($type, 0, -2)))->isInstance($object);
   } else if (0 === substr_compare($type, '[:', 0, 2)) {
     return (new \lang\MapType(substr($type, 2, -1)))->isInstance($object);
-  } else if (0 === strncmp($type, 'function(', 9)) {
-    return \lang\FunctionType::forName($type)->isInstance($object);
+  } else if (0 === strncmp($type, '(function(', 10)) {
+    return \lang\FunctionType::forName(substr($type, 1, -1))->isInstance($object);
   } else if (strstr($type, '?')) {
     return \lang\WildcardType::forName($type)->isInstance($object);
   } else {